requirements.txt
opencv
pillow
pyautogui